+++ src/main/java/see/hm/edu/handlers/CancelandStopIntentHandler.java
@@ -11,7 +11,7 @@
      the specific language governing permissions and limitations under the License.
 */
 
-package main.java.quiz.handlers;
+package see.hm.edu.handlers;
 
 import com.amazon.ask.dispatcher.request.handler.HandlerInput;
 import com.amazon.ask.dispatcher.request.handler.RequestHandler;
@@ -21,30 +21,17 @@ import java.util.Optional;
 
 import static com.amazon.ask.request.Predicates.intentName;
 
-public class WhatsMyColorIntentHandler implements RequestHandler {
-    public static final String COLOR_KEY = "COLOR";
-    public static final String COLOR_SLOT = "Color";
-
+public class CancelandStopIntentHandler implements RequestHandler {
     @Override
     public boolean canHandle(HandlerInput input) {
-        return input.matches(intentName("WhatsMyColorIntent"));
+        return input.matches(intentName("AMAZON.StopIntent").or(intentName("AMAZON.CancelIntent")));
     }
 
     @Override
     public Optional<Response> handle(HandlerInput input) {
-        String speechText;
-        String favoriteColor = (String) input.getAttributesManager().getSessionAttributes().get(COLOR_KEY);
-
-        if (favoriteColor != null && !favoriteColor.isEmpty()) {
-            speechText = String.format("Deine Lieblingsfarbe ist %s. Auf Wiedersehen.", favoriteColor);
-        } else {
-            // Since the user's favorite color is not set render an error message.
-            speechText = "Ich weiss nicht welches Deine Lieblingsfarbe ist. Sag mir Deine Lieblingsfarbe. Sage zum Beispiel: ich mag rot.";
-        }
-
         return input.getResponseBuilder()
-                .withSpeech(speechText)
-                .withSimpleCard("ColorSession", speechText)
+                .withSpeech("Auf Wiedersehen")
+                .withSimpleCard("ColorSession", "Auf Wiedersehen")
                 .build();
     }
 }
