+++ src/main/java/see/hm/edu/handlers/DefinePlayersIntentHandler.java
@@ -11,7 +11,7 @@
      the specific language governing permissions and limitations under the License.
 */
 
-package main.java.colorpicker.handlers;
+package see.hm.edu.handlers;
 
 import com.amazon.ask.dispatcher.request.handler.HandlerInput;
 import com.amazon.ask.dispatcher.request.handler.RequestHandler;
@@ -21,29 +21,23 @@ import com.amazon.ask.model.Request;
 import com.amazon.ask.model.Response;
 import com.amazon.ask.model.Slot;
 import com.amazon.ask.response.ResponseBuilder;
-import com.opencsv.CSVReader;
 
-import main.java.colorpicker.ColorPickerStreamHandler;
+import see.hm.edu.quiz.ColorPickerStreamHandler;
 
-import java.io.FileReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
 import static com.amazon.ask.request.Predicates.intentName;
 
-public class AskQuestionIntentHandler implements RequestHandler {
+public class DefinePlayersIntentHandler implements RequestHandler {
     @Override
     public boolean canHandle(HandlerInput input) {
-        return input.matches(intentName("AskQuestionIntent"));
+        return input.matches(intentName("DefinePlayersIntent"));
     }
 
     @Override
     public Optional<Response> handle(HandlerInput input) {
-    	/*
         Request request = input.getRequestEnvelope().getRequest();
         IntentRequest intentRequest = (IntentRequest) request;
         Intent intent = intentRequest.getIntent();
@@ -80,35 +74,22 @@ public class AskQuestionIntentHandler implements RequestHandler {
 
         } else {
             // Render an error since we don't know what the users favorite color is.
-            speechText = "Ungige Eingabe. Es knnen zwischen 1 und 4 Menschen mitspielen. Bitte whle erneut aus";
+            speechText = "Ungültige Eingabe. Es können zwischen 1 und 4 Menschen mitspielen. Bitte wähle erneut aus";
             repromptText =
                     "Ich weiss nicht welches Deine Lieblingsfarbe ist. Sag mir Deine Lieblingsfarbe. Sage zum Beispiel: ich mag blau.";
             isAskResponse = true;
         }
-	*/
-    	String speechText = "test";
-    	try {
-    		
-    	 CSVReader reader = new CSVReader(new InputStreamReader(AskQuestionIntentHandler.class.getResourceAsStream("/Bsp.csv")), ',', '"', 1);
-	       
-	      //Read all rows at once
-	      List<String[]> allRows = reader.readAll();
-	      
-	      String[] strings = allRows.get(0);
-	      speechText = strings[1];
-    	}
-    	catch(IOException e) {} 
-    	
+
         ResponseBuilder responseBuilder = input.getResponseBuilder();
 
         responseBuilder.withSimpleCard("ColorSession", speechText)
                 .withSpeech(speechText)
                 .withShouldEndSession(false);
 
-     /*   if (isAskResponse) {
+        if (isAskResponse) {
             responseBuilder.withShouldEndSession(false)
                     .withReprompt(repromptText);
-        } */
+        }
 
         return responseBuilder.build();
     }
