+++ src/main/java/see/hm/edu/handlers/FallbackIntentHandler.java
@@ -1,4 +1,4 @@
-package main.java.quiz.handlers;
+package see.hm.edu.handlers;
 
 import com.amazon.ask.dispatcher.request.handler.HandlerInput;
 import com.amazon.ask.dispatcher.request.handler.RequestHandler;
