+++ src/main/java/see/hm/edu/quiz/ColorPickerStreamHandler.java
@@ -11,21 +11,13 @@
      the specific language governing permissions and limitations under the License.
 */
 
-package main.java.colorpicker;
+package see.hm.edu.quiz;
 
 import com.amazon.ask.Skill;
 import com.amazon.ask.SkillStreamHandler;
 import com.amazon.ask.Skills;
-import main.java.quiz.handlers.FallbackIntentHandler;
-import main.java.quiz.handlers.HelpIntentHandler;
-import main.java.quiz.handlers.LaunchRequestHandler;
-import main.java.quiz.handlers.SessionEndedRequestHandler;
-import main.java.quiz.handlers.WhatsMyColorIntentHandler;
-import main.java.colorpicker.handlers.AskQuestionIntentHandler;
-import main.java.quiz.handlers.CancelandStopIntentHandler;
-import main.java.quiz.handlers.MyColorIsIntentHandler;
-import main.java.quiz.handlers.StartQuizIntentHandler;
-import main.java.quiz.handlers.DefinePlayersIntentHandler;
+
+import see.hm.edu.handlers.*;
 
 public class ColorPickerStreamHandler extends SkillStreamHandler {
 	public static int num_players;
@@ -42,7 +34,6 @@ public class ColorPickerStreamHandler extends SkillStreamHandler {
                         new SessionEndedRequestHandler(),
                         new HelpIntentHandler(),
                         new DefinePlayersIntentHandler(),
-                        new AskQuestionIntentHandler(),
                         new FallbackIntentHandler())
                 // Add your skill id below
                 //.withSkillId("")
