+++ src/main/java/see/hm/edu/quizmodel/Answer.java
@@ -1,4 +1,4 @@
-package main.java.quiz.model;
+package see.hm.edu.quizmodel;
 
 public class Answer {
 	public String text;
