+++ src/main/java/see/hm/edu/quizmodel/QuizRound.java
@@ -1,4 +1,4 @@
-package main.java.quiz.model;
+package see.hm.edu.quizmodel;
 
 import java.util.Arrays;
