+++ src/main/java/see/hm/edu/quizmodel/Region.java
@@ -1,4 +1,4 @@
-package main.java.quiz.model;
+package see.hm.edu.quizmodel;
 
 import java.net.URL;
